fix(ci): include EmpoLogic helper sources for Linux tests
